# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Origin Agent Platform Product API
  version: v1.0-alpha
  description: >
    Product-level API contract for Origin Agent Platform. The product owns this
    contract; server-side services implement it; PloyKit modules consume it.
...
servers:
  - url: https://api.origin.example
    description: Production API endpoint.
  - url: http://127.0.0.1:8080
    description: Local implementation endpoint.
  - url: http://127.0.0.1:4010
    description: Local contract/mock endpoint.
security:
  - serviceToken: []
    originSignature: []
    originTimestamp: []
    originClaims: []
...
paths:
...
  /v1/agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: List agents.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Agent list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
    post:
      tags: [Agents]
      operationId: createAgent
      summary: Create an agent.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreateRequest"
      responses:
        "201":
          description: Agent created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /v1/agents/{agent_id}:
    parameters:
      - $ref: "#/components/parameters/AgentId"
    get:
      tags: [Agents]
      operationId: getAgent
      summary: Get an agent.
      responses:
        "200":
          description: Agent resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Agents]
      operationId: patchAgent
      summary: Patch an agent.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentPatchRequest"
      responses:
        "200":
          description: Agent patched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Agents]
      operationId: deleteAgent
      summary: Delete an agent.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Agent deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/agents/{agent_id}/collaborators:
    patch:
      tags: [Agents]
      operationId: patchAgentCollaborators
      summary: Update agent collaborators and access roles.
      parameters:
        - $ref: "#/components/parameters/AgentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCollaboratorsPatchRequest"
      responses:
        "200":
          description: Agent collaborators updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetailsResponse"
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: service-token
    originSignature:
      type: apiKey
      in: header
      name: x-origin-agentops-signature
    originTimestamp:
      type: apiKey
      in: header
      name: x-origin-agentops-timestamp
    originClaims:
      type: apiKey
      in: header
      name: x-origin-agentops-claims
  parameters:
...
    IdempotencyKey:
      name: idempotency-key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Conflict, usually idempotency or state transition related.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Meta:
      type: object
      required: [request_id, trace_id]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        workspace_id:
          type: string
        tenant_id:
          type: string
      additionalProperties: true
    ListMeta:
      allOf:
        - $ref: "#/components/schemas/Meta"
        - type: object
          properties:
            next_cursor:
              type: string
              nullable: true
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          $ref: "#/components/schemas/Meta"
...
    ResourceRef:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
    Agent:
      type: object
      required: [agent_id, name, status, created_at, updated_at]
      properties:
        agent_id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, validating, review_required, published, disabled, archived]
        version:
          type: string
        identifier:
          type: string
          description: Human-friendly stable slug used in product UI.
        workspace_id:
          type: string
        version_policy:
          type: string
        prompt:
          type: string
        prompt_profile:
          type: object
          additionalProperties: true
        skill_ids:
          type: array
          items:
            type: string
        tool_ids:
          type: array
          items:
            type: string
        provider_ids:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        objectives:
          type: array
          items:
            type: string
        draft_version:
          type: string
          nullable: true
        last_run_at:
          type: string
          format: date-time
          nullable: true
        last_run_status:
          type: string
          nullable: true
        tool_count:
          type: integer
        run_count:
          type: integer
        eval_summary:
          type: object
          additionalProperties: true
        governance_summary:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        identifier:
          type: string
        workspace_id:
          type: string
        version_policy:
          type: string
        skill_ids:
          type: array
          items:
            type: string
        tool_ids:
          type: array
          items:
            type: string
        provider_ids:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        objectives:
          type: array
          items:
            type: string
        approval_required:
          type: boolean
        external_api_allowed:
          type: boolean
        tracing_enabled:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
    AgentPatchRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        identifier:
          type: string
        workspace_id:
          type: string
        version_policy:
          type: string
        status:
          type: string
        skill_ids:
          type: array
          items:
            type: string
        tool_ids:
          type: array
          items:
            type: string
        provider_ids:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        objectives:
          type: array
          items:
            type: string
        approval_required:
          type: boolean
        external_api_allowed:
          type: boolean
        tracing_enabled:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
...
    AgentResponse:
      type: object
      required: [agent, meta]
      properties:
        agent:
          $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/Meta"
    AgentListResponse:
      type: object
      required: [agents, meta]
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/ListMeta"
    Skill:
      type: object
      required: [skill_id, name, status]
      properties:
        skill_id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, enabled, disabled, archived]
        kind:
          type: string
          enum: [flow, atomic, composite]
        version:
          type: string
        tags:
          type: array
          items:
            type: string
        instruction:
          type: string
        input_schema:
          type: object
          additionalProperties: true
        output_schema:
          type: object
          additionalProperties: true
        resources:
          type: array
          items:
            $ref: "#/components/schemas/ResourceRef"
        usage_count:
          type: integer
        success_rate:
          type: number
          format: double
        value_score:
          type: integer
        agent_refs:
          type: array
          items:
            type: string
        creator:
          type: string
        updater:
          type: string
        last_used_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
...
    Tool:
      type: object
      required: [tool_id, name, provider_id, status]
      properties:
        tool_id:
          type: string
        name:
          type: string
        description:
          type: string
        provider_id:
          type: string
        group_id:
          type: string
        status:
          type: string
          enum: [draft, active, available, limited, unavailable, disabled, archived]
        version:
          type: string
        method:
          type: string
        path:
          type: string
        tool_type:
          type: string
        input_schema:
          type: object
          additionalProperties: true
        output_schema:
          type: object
          additionalProperties: true
        risk_level:
          type: string
          enum: [low, medium, high, critical]
        tags:
          type: array
          items:
            type: string
        usage_count:
          type: integer
        success_rate:
          type: number
          format: double
        failure_count:
          type: integer
        input_count:
          type: integer
        output_count:
          type: integer
        last_smoke_status:
          type: string
          nullable: true
        last_smoke_at:
          type: string
          format: date-time
          nullable: true
        creator:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    Run:
      type: object
      required: [run_id, agent_id, status, objective, created_at]
      properties:
        run_id:
          type: string
        agent_id:
          type: string
        agent_name:
          type: string
        agent_version:
          type: string
        task_id:
          type: string
        trace_id:
          type: string
        status:
          type: string
          enum: [queued, running, waiting_input, waiting_approval, approval_required, succeeded, failed, cancelled]
        trigger:
          type: string
        environment:
          type: string
        user:
          type: string
        objective:
          type: string
        tool_calls:
          type: integer
        tokens:
          type: integer
        duration_ms:
          type: integer
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        artifact_refs:
          type: array
          items:
            type: string
        result:
          type: object
          additionalProperties: true
        error:
          $ref: "#/components/schemas/ErrorEnvelope"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
...
    Approval:
      type: object
      required: [approval_id, action, status, resource, created_at]
      properties:
        approval_id:
          type: string
        action:
          type: string
        approval_type:
          type: string
        status:
          type: string
          enum: [pending, waiting_approval, waiting_info, approved, rejected, cancelled, expired, resolved]
        risk:
          type: string
          enum: [low, medium, high, critical]
        resource:
          $ref: "#/components/schemas/ResourceRef"
        agent_id:
          type: string
        agent_name:
          type: string
        run_id:
          type: string
        source:
          type: string
        environment:
          type: string
        reason:
          type: string
        impact:
          type: array
          items:
            type: string
        request_payload:
          type: object
          additionalProperties: true
        approval_chain:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalStep"
        sla_due_at:
          type: string
          format: date-time
        requested_by:
          type: string
        resolved_by:
          type: string
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
...
    TimeSeriesPoint:
      type: object
      required: [timestamp, value]
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
          format: double
        label:
          type: string
    TimelineEvent:
      type: object
      required: [event_id, title, occurred_at]
      properties:
        event_id:
          type: string
        title:
          type: string
        status:
          type: string
        category:
          type: string
        occurred_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
        actor:
          type: string
        payload:
          type: object
          additionalProperties: true
...
    AgentUsageSummary:
      type: object
      properties:
        total_runs:
          type: integer
        successful_runs:
          type: integer
        failed_runs:
          type: integer
        success_rate:
          type: number
          format: double
        tokens_used:
          type: integer
        token_limit:
          type: integer
        tool_calls:
          type: integer
        tool_call_limit:
          type: integer
        run_series:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
    AgentVersion:
      type: object
      required: [version, status, created_at]
      properties:
        version:
          type: string
        status:
          type: string
          enum: [draft, canary, stable, archived]
        release_notes:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          additionalProperties: true
    AgentCollaborator:
      type: object
      required: [member_id, role]
      properties:
        member_id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, editor, viewer]
        avatar_url:
          type: string
    AgentDetailsResponse:
      type: object
      required: [agent, meta]
      properties:
        agent:
          $ref: "#/components/schemas/Agent"
        usage:
          $ref: "#/components/schemas/AgentUsageSummary"
        tool_bindings:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        runs:
          type: array
          items:
            $ref: "#/components/schemas/Run"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/Approval"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/AgentVersion"
        collaborators:
          type: array
          items:
            $ref: "#/components/schemas/AgentCollaborator"
        access_control:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/Meta"
...
    AgentCollaboratorsPatchRequest:
      type: object
      required: [collaborators]
      properties:
        collaborators:
          type: array
          items:
            $ref: "#/components/schemas/AgentCollaborator"
...
    ApprovalStep:
      type: object
      required: [step_id, role, status]
      properties:
        step_id:
          type: string
        role:
          type: string
        member_id:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, skipped]
        acted_at:
          type: string
          format: date-time
          nullable: true
...